feat: rename home screen to calendar screen and add floating action button
